# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NuweFarma API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/proveedores": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Listar proveedores",
                "operationId": "7f6ed20dc86f1682aa7e31d71afa37df",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Buscar por nombre, NIT, email o contacto",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "inactivo"
                            ]
                        }
                    },
                    {
                        "name": "ciudad",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de proveedores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Crear proveedor",
                "operationId": "a1046f3d7a3c804651bf420761995e57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/proveedores/{id}": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Ver proveedor",
                "operationId": "5aa647e1f6ead15a2c088ecc1fbb12bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de proveedor"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Actualizar proveedor",
                "operationId": "28ff17fc2a31c61745cbe9a18da5aa32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Eliminar proveedor",
                "operationId": "2b9dc197602ee96f15138b4ad2024fab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/proveedores/{id}/toggle-estado": {
            "patch": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Activar/Desactivar proveedor",
                "operationId": "73a73f506d6f026f847be66548783bd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}